move all this data fetching into a component
