+++ app/components/fetch-results/template.hbs
@@ -1,0 +1,6 @@
+{{yield (hash
+  ny=ny
+  nj=nj
+  swing=swing
+  race=race
+)}}
